+++ js/data/game-store.js
@@ -8,42 +8,46 @@ class GameStore {
       time: 50
     });
     this.currentState = Object.assign({}, this.initialState);
-    this.guessArtistData = {
-      src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
-      answers: [
-        {
-          text: `Пелагея`,
-          image: `https://i.vimeocdn.com/portrait/992615_300x300`
-        },
-        {
-          text: `Краснознаменная дивизия имени моей бабушки`,
-          image: `https://i.ytimg.com/vi/jzgM3m8Vp1k/maxresdefault.jpg`
-        },
-        {
-          text: `Lorde`,
-          image: `https://yt3.ggpht.com/-fkDeGauT7Co/AAAAAAAAAAI/AAAAAAAAAAA/dkF5ZKkrxRo/s900-c-k-no-mo-rj-c0xffffff/photo.jpg`
-        }
-      ],
-      rightAnswer: 2
-    };
-    this.guessGenreData = {
-      text: `Выберите инди-рок треки`,
-      answers: [
-        {
-          src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
-        },
+    this.data = {
+      guessArtistData: {
+        type: `artist`,
+        src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
+        answers: [
+          {
+            text: `Пелагея`,
+            image: `https://i.vimeocdn.com/portrait/992615_300x300`
+          },
+          {
+            text: `Краснознаменная дивизия имени моей бабушки`,
+            image: `https://i.ytimg.com/vi/jzgM3m8Vp1k/maxresdefault.jpg`
+          },
+          {
+            text: `Lorde`,
+            image: `https://yt3.ggpht.com/-fkDeGauT7Co/AAAAAAAAAAI/AAAAAAAAAAA/dkF5ZKkrxRo/s900-c-k-no-mo-rj-c0xffffff/photo.jpg`
+          }
+        ],
+        rightAnswer: 2
+      },
+      guessGenreData: {
+        type: `genre`,
+        text: `Выберите инди-рок треки`,
+        answers: [
+          {
+            src: `https://www.youtube.com/audiolibrary_download?vid=91624fdc22fc54ed`,
+          },
 
-        {
-          src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
-        },
-        {
-          src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
-        },
-        {
-          src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
-        }
-      ],
-      rightAnswer: [4]
+          {
+            src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
+          },
+          {
+            src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
+          },
+          {
+            src: `https://www.youtube.com/audiolibrary_download?vid=bcbe5be936a32fb1`,
+          }
+        ],
+        rightAnswer: [4]
+      }
     };
   }
